Performance test program
This patch adds a new test program that checks the performance of
various operations in the Avro library.  Right now the only thing we
check is the speed of our reference counting implementation.  The new
test case isn't run as part of "make check", since if we're doing our
jobs right, it'll take quite awhile to run.