Atomic reference counts
This patch updates all of the reference counts in various Avro objects
to be updated atomically.  This takes care of some race conditions when
Avro instances are passed between threads.  The atomic increment and
decrement operations are slower than the previous non-atomic operations,
but according to our performance tests, reference counting isn't the
dominating factor in the running time.