Resizable array class
This patch adds a new helper class for holding a resizable array of
fixed-length data.  We'll use this in the implementation of the
schema-specific class for Avro array types.  We can also use it to
replace the numerically-indexed hash tables that are used in several of
the avro_schema_t and avro_datum_t subclasses.