Resizable map class
This patch adds a new helper class for holding a map of fixed-length
data, where the elements can be retrieved both by their string key, and
by numeric index.  (The numeric indices are assigned based on the order
that the elements are added to the map.)  We'll use this in the
implementation of the schema-specific class for Avro map types.