Schema-specific structs
This patch adds a new avrocc command, which produces a C struct that is
specific to an Avro schema.  Right now, there's not especially much that
you can do with that schema; there are lifecycle functions for creating
and destroying instances of those types, and that's it.

The compiler works by producing a “spec” file, which contains an
equivalent definition of the schema using C preprocessor macros.  It
also produces a skeleton C header and implementation file.  The actual
schema-specific C definitions are then created using the C preprocessor
as the template language.  Each different type or function requires a
separate preprocessor file, which provides its own definitions of the
macros used in the spec file.  In this way, the spec file drives the C
preprocessor to produce the appropriate definitions for the schema.