Schema-specific union functions
This patch adds a handful of functions for handling the schema-specific
types that we generate for union schemas.  There's a _clear() function,
which resets the union instance to not point at any branch.  There are
also _set_foo() functions for each branch in the union; these ensure
that the union instance current contains an instance of the given
branch, initializing the branch object if necessary.  If the union
already contains that branch, nothing happens.