Child consumers in base consumer class
Before, the resolver subclass maintained a list of “child” consumers,
which would be used to store the resolvers for each field in a record,
element in an array, etc.  While working on the schema-specific
resolvers, it turned out that this is a common need, so we've moved this
array into the base consumer class.  The avro_consumer_free() function
takes care of recursively freeing any child consumers in the array.  We
allow consumer cycles in the array, so that we can support recursive
schemas easily; the avro_consumer_free() function also takes care of
ensuring that each consumer in a cycle is freed only once.