Public memoization API
The resolver class memoizes the results of its schema resolution
algorithm, which allows us to support recursive schemas; if we didn't do
this, we'd get an infinite loop, since we'd repeatedly drill down into
the recursive cycle of schema references while trying to resolve child
schemas.  In turns out we'll need this behavior in the schema-specific
resolvers, too, so this patch extracts the memoization behavior out into
a separate public API.  For now, it's hard-coded to use two pointer
parameters as the memoization key.  We can expand this if needed, but it
serves our purposes for now.