Resolver class for schema-specific reader unions
This patch adds a resolver implementation for the schema-specific type
that we generate for union schemas.  This implementation only applies
when the _reader_ schema is a union.  The major difference is that we
need to add extra logic to each of the existing resolver callbacks, to
select the appropriate branch of the reader value when the reader schema
is a union.  This is a cross-cutting change, since every resolver class
needs to perform this check (even the predefined resolvers for the raw
primitive types).  We represent the union branch with a pointer to one
of the schema-specific “set branch” functions.  If this pointer is null,
then the target is a bare value, and not a branch within a union.  If
it's non-NULL, then the user_data pointer passed into the callback
points at the union value, and we use the function to select the branch.