Resolver class for schema-specific writer unions
This patch adds a new class for reading instances of a writer union into
a schema-specific type.  For writer unions, we only need a single class;
we don't need to generate a new one on the fly for each schema-specific
type.  The writer union constructor takes in a pointer to a function
that can try to resolve each branch of the writer union against the
schema-specific type.  This gives us an array of child consumers, one
for each branch of the writer union.  When reading in the data, we use
the writer union discriminant to choose the appropriate child consumer,
and pass off control to it.