avrocc works on multiple input schema files
The avrocc compiler now accepts multiple .avsc schema files.  It still
produces a single set of output files, which will contain all of the
schemas and subschemas defined in the .avsc files that are passed in.
We make sure to only output any given type once, even if it appears in
multiple .avsc files.