ResolvedReader/Writer don't maintain references
The ResolvedReader and Writer classes use to increment the reference
count of whatever value you passed in to set_source and set_dest, in an
attempt to ensure that the source/destination value stays around for at
least as long as the resolved value.  Unfortunately, this means that you
can't use a child value as the source or destination of a resolver,
since child values don't have valid incref and decref implementations.

Now, you're fully responsible for managing the lifecycle of all of the
values involved with schema resolution.