Hide Jansson symbols in Avro library
The Avro C library embeds a copy of the Jansson library for handling
JSON values.  Unfortunately, the build scripts include all of the
Jansson symbols in the publically exported API of libavro.so.  That
gives us conflicts if we try to link with both libavro and a separately
installed copy of libjansson.so.  This patch updates our embedded copy
of Jansson to hide its symbols, so that they're only visible within
libavro.so.